Move controller status history table cells into entityStatus directory
