+++ src/components/tables/cells/entityStatus/ControllerStatus/types.ts
@@ -1,14 +1,15 @@
-import { PopperPlacementType } from '@mui/material';
+import { Error } from 'deps/control-plane/types';
 import { JobStatus } from 'types';
+import { StatusColor } from 'utils/entityStatus-utils';
 
 export interface ControllerStatusProps {
     detail: string;
     status?: JobStatus;
 }
 
-export interface DetailProps {
-    detail: string;
-    popperPlacement?: PopperPlacementType;
+export interface ChipLabelProps {
+    color: StatusColor;
+    errors: Error[];
 }
 
 export interface StatusIndicatorProps {
