+++ src/components/tables/cells/entityStatus/types.ts
@@ -1,0 +1,6 @@
+import { PopperPlacementType } from '@mui/material';
+
+export interface ControllerEventProps {
+    detail: string;
+    popperPlacement?: PopperPlacementType;
+}
