+++ src/utils/entityStatus-utils.ts
@@ -7,24 +7,38 @@ import {
 } from 'context/Theme';
 import { JobStatus } from 'deps/control-plane/types';
 
+type MuiColorId =
+    | 'default'
+    | 'primary'
+    | 'secondary'
+    | 'error'
+    | 'info'
+    | 'success'
+    | 'warning';
+
 // The hex string additions correspond to sample_grey[500] | sample_grey[300].
-export type StatusColor = SemanticColor | '#C4D3E9' | '#E1E9F4';
+export type StatusColorHex = SemanticColor | '#C4D3E9' | '#E1E9F4';
+
+export interface StatusColor {
+    hex: StatusColorHex;
+    id: MuiColorId;
+}
 
 export const getStatusIndicatorColor = (
     colorMode: PaletteMode,
     status?: JobStatus
 ): StatusColor => {
     if (status?.type === 'success' || status?.type === 'emptyDraft') {
-        return successMain;
+        return { hex: successMain, id: 'success' };
     }
 
     if (status?.type === 'queued') {
-        return warningMain;
+        return { hex: warningMain, id: 'warning' };
     }
 
     if (status?.type.includes('Fail')) {
-        return errorMain;
+        return { hex: errorMain, id: 'error' };
     }
 
-    return colorMode === 'dark' ? '#E1E9F4' : '#C4D3E9';
+    return { hex: colorMode === 'dark' ? '#E1E9F4' : '#C4D3E9', id: 'default' };
 };
